Automatically backed up by Learn

+++ key_for_min.rb
@@ -1,6 +1,6 @@
 def key_for_min_value(name_hash)
   min_key = nil
-  min_value = Integer::INFINITY
+  min_value = Float::INFINITY
   name_hash.each do |key, value|
     if value < min_value
       min_value = value
